Rework Spell Data to 'modules contain all' format
