+++ src/Main/MonkSpreadsheet.js
@@ -74,15 +74,15 @@ class MonkSpreadsheet extends React.Component {
           <tr style={styles.cellBorder}><td>{parser.modules.upliftingTrance.UTProcsTotal}</td></tr>
           <tr><td>{(1 - (parser.modules.upliftingTrance.consumedUTProc / parser.modules.upliftingTrance.UTProcsTotal)).toFixed(4)}</td></tr>
           <tr style={styles.cellBorder}><td>{(parser.modules.manaTea.manaSavedMT / parser.fightDuration * 1000 * 5).toFixed(0)}</td></tr>
-          <tr style={styles.cellBorder}><td>{(parser.selectedCombatant.hasTalent(SPELLS.LIFECYCLES_TALENT.id) && ((parser.modules.manaSavingTalents.castsRedViv / (parser.modules.manaSavingTalents.castsRedViv + parser.modules.manaSavingTalents.castsNonRedViv)).toFixed(4))) || 0}</td></tr>
-          <tr><td>{(parser.selectedCombatant.hasTalent(SPELLS.LIFECYCLES_TALENT.id) && ((parser.modules.manaSavingTalents.castsRedEnm / (parser.modules.manaSavingTalents.castsRedEnm + parser.modules.manaSavingTalents.castsNonRedEnm)).toFixed(4))) || 0}</td></tr>
-          <tr style={styles.cellBorder}><td>{parser.modules.manaSavingTalents.manaReturnSotc}</td></tr>
+          <tr style={styles.cellBorder}><td>{(parser.selectedCombatant.hasTalent(SPELLS.LIFECYCLES_TALENT.id) && ((parser.modules.lifecycles.castsRedViv / (parser.modules.lifecycles.castsRedViv + parser.modules.lifecycles.castsNonRedViv)).toFixed(4))) || 0}</td></tr>
+          <tr><td>{(parser.selectedCombatant.hasTalent(SPELLS.LIFECYCLES_TALENT.id) && ((parser.modules.lifecycles.castsRedEnm / (parser.modules.lifecycles.castsRedEnm + parser.modules.lifecycles.castsNonRedEnm)).toFixed(4))) || 0}</td></tr>
+          <tr style={styles.cellBorder}><td>{parser.modules.spiritOfTheCrane.manaReturnSotc}</td></tr>
           <tr style={styles.cellBorder}><td>{(parser.modules.sheilunsGift.stacksTotalSG / SGcasts || 0).toFixed(0)}</td></tr>
           <tr style={styles.cellBorder}><td>{(parser.modules.sheilunsGift.countWhispersHeal / ((Math.floor(parser.fightDuration / 10000) + parser.modules.sheilunsGift.countEff))).toFixed(4) || 0}</td></tr>
-          <tr style={styles.cellBorder}><td>{((parser.modules.aoeHealingTracker.healsCelestialBreath / parser.modules.aoeHealingTracker.procsCelestialBreath) / 3).toFixed(2) || 0}</td></tr>
-          <tr style={styles.cellBorder}><td>{parser.modules.aoeHealingTracker.procsMistsOfSheilun}</td></tr>
-          <tr><td>{(parser.modules.aoeHealingTracker.healsMistsOfSheilun / parser.modules.aoeHealingTracker.procsMistsOfSheilun).toFixed(2) || 0}</td></tr>
-          <tr style={styles.cellBorder}><td>{(parser.selectedCombatant.hasTalent(SPELLS.REFRESHING_JADE_WIND_TALENT.id) && ((parser.modules.aoeHealingTracker.healsRJW / parser.modules.aoeHealingTracker.castRJW) / 78 ).toFixed(4)) || 0}</td></tr>
+          <tr style={styles.cellBorder}><td>{((parser.modules.celestialBreath.healsCelestialBreath / parser.modules.celestialBreath.procsCelestialBreath) / 3).toFixed(2) || 0}</td></tr>
+          <tr style={styles.cellBorder}><td>{parser.modules.mistsOfSheilun.procsMistsOfSheilun}</td></tr>
+          <tr><td>{(parser.modules.mistsOfSheilun.healsMistsOfSheilun / parser.modules.mistsOfSheilun.procsMistsOfSheilun).toFixed(2) || 0}</td></tr>
+          <tr style={styles.cellBorder}><td>{(parser.selectedCombatant.hasTalent(SPELLS.REFRESHING_JADE_WIND_TALENT.id) && ((parser.modules.refreshingJadeWind.healsRJW / parser.modules.refreshingJadeWind.castRJW) / 78 ).toFixed(4)) || 0}</td></tr>
           <tr style={styles.cellBorder}><td>{parser.modules.renewingMist.dancingMistHeal}</td></tr>
           <tr style={styles.cellBorder}><td>{(((parser.modules.essenceFontMastery.healEF / 2) / parser.modules.essenceFontMastery.castEF) || 0).toFixed(2)}</td></tr>
           <tr style={styles.cellBorder}><td>{(parser.modules.essenceFontMastery.targetsEF / parser.modules.essenceFontMastery.castEF).toFixed(2) || 0}</td></tr>
