+++ src/Parser/MistweaverMonk/Modules/Spells/UpliftingTrance.js
@@ -1,21 +1,25 @@
 // Based on Clearcasting Implementation done by @Blazyb
+import React from 'react';
+
+import SPELLS from 'common/SPELLS';
+import SpellLink from 'common/SpellLink';
+import SpellIcon from 'common/SpellIcon';
+import { formatPercentage } from 'common/format';
 
 import Module from 'Parser/Core/Module';
-import SPELLS from 'common/SPELLS';
+
+import StatisticBox, { STATISTIC_ORDER } from 'Main/StatisticBox';
 
 const UT_DURATION = 20000;
 const debug = false;
 
 class UpliftingTrance extends Module {
-
   UTProcsTotal = 0;
   lastUTProcTime = 0;
   consumedUTProc = 0;
   overwrittenUTProc = 0;
   nonUTVivify = 0;
-
   tftVivCast = 0;
-
 
   on_byPlayer_applybuff(event) {
     const spellId = event.ability.guid;
@@ -59,6 +63,34 @@ class UpliftingTrance extends Module {
       }
     }
   }
+
+  suggestions(when) {
+    const unusedUTProcs = 1 - (this.consumedUTProc / this.UTProcsTotal);
+    when(unusedUTProcs).isGreaterThan(.3)
+      .addSuggestion((suggest, actual, recommended) => {
+        return suggest(<span>Your <SpellLink id={SPELLS.UPLIFTING_TRANCE_BUFF.id} /> procs should be used as soon as you get them so they are not overwritten. While some will be overwritten due to the nature of the spell interactions, holding <SpellLink id={SPELLS.UPLIFTING_TRANCE_BUFF.id} /> procs is not optimal.</span>)
+          .icon(SPELLS.UPLIFTING_TRANCE_BUFF.icon)
+          .actual(`${formatPercentage(unusedUTProcs)}% Unused Uplifting Trance procs`)
+          .recommended(`<${formatPercentage(recommended)}% wasted UT Buffs is recommended`)
+          .regular(recommended + .1).major(recommended + .2);
+    });
+  }
+
+  statistic() {
+    const unusedUTProcs = 1 - (this.consumedUTProc / this.UTProcsTotal);
+    return (
+      <StatisticBox
+        icon={<SpellIcon id={SPELLS.UPLIFTING_TRANCE_BUFF.id} />}
+        value={`${formatPercentage(unusedUTProcs)}%`}
+        label={(
+          <dfn data-tip={`You got total <b>${this.UTProcsTotal} uplifting trance procs</b> and <b>used ${this.consumedUTProc}</b> of them. ${this.nonUTVivify} of your vivify's were used without an uplifting trance procs.`}>
+            Unused Procs
+          </dfn>
+        )}
+      />
+    );
+  }
+  statisticOrder = STATISTIC_ORDER.OPTIONAL();
 }
 
 export default UpliftingTrance;
